pyproject.toml
[tool.poetry]
name = "aplicativos-fiscal"
version = "0.1.0"
description = ""
authors = ["Rodolfo Scarp <rodolfo@jrcontabiltr.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
openpyxl = "^3.1.2"
pyinstaller = "^6.6.0"
pyside6 = "^6.7.0"
firebirdsql = "^1.2.5"
cryptography = "^42.0.5"
loguru = "^0.7.2"
pandas = "^2.2.2"
pydantic = "^2.7.1"
validate-docbr = "^1.10.0"
pyperclip = "^1.8.2"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "^8.2.0"
pylint = "^3.1.0"
taskipy = "^1.12.2"
isort = "^5.13.2"
blue = "^0.9.1"

[tool.pytest.ini_options]
addopts = ["--doctest-modules", "-vv", "-x"]

[tool.poetry.group.docs.dependencies]
mkdocs = { extras = ["i18n"], version = "^1.6.0" }
mkdocs-material = "^9.5.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
test = "pytest --pyargs src/core"
docs = "mkdocs serve"
format = "isort src && blue src"
